Extracting ClientForBulkAPIs
